[hw10 - task1] consider annotations at generating

+++ src/main/java/edu/hw10/Task1/annotations/Min.java
@@ -9,6 +9,6 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 public @interface Min {
 
-    int value();
+    double value();
 
 }
